Extract non-proto annotator
Also:
  * Rename `Annotator` to `ProtoAnnotator`.
