+++ mc-java-annotation/src/main/kotlin/io/spine/tools/mc/java/annotation/AbstractAnnotator.kt
@@ -26,11 +26,10 @@
 
 package io.spine.tools.mc.java.annotation
 
-import io.spine.base.EntityState
 import io.spine.protodata.codegen.java.JavaRenderer
 import io.spine.protodata.renderer.SourceFileSet
-import io.spine.tools.mc.annotation.ApiAnnotationsPlugin
 import io.spine.protodata.settings.loadSettings
+import io.spine.tools.mc.annotation.ApiAnnotationsPlugin
 import io.spine.tools.mc.annotation.ApiOption
 import io.spine.tools.mc.annotation.ApiOption.BETA
 import io.spine.tools.mc.annotation.ApiOption.EXPERIMENTAL
@@ -38,18 +37,13 @@ import io.spine.tools.mc.annotation.ApiOption.INTERNAL
 import io.spine.tools.mc.annotation.ApiOption.SPI
 
 /**
- * An abstract base for annotation renderers.
- *
- * @param T the type of the view state which contains information about annotated types.
+ * The abstract base for renderers that annotate Java source code using
+ * the [annotation types][annotationClass] passed via an instance of [Settings].
  */
-internal abstract class Annotator<T>(
-    private val viewClass: Class<T>
-) : JavaRenderer() where T : EntityState<*> {
-
-    protected lateinit var sources: SourceFileSet
+internal abstract class AbstractAnnotator: JavaRenderer() {
 
     /**
-     * Returns cannonical name of [ApiAnnotationsPlugin].
+     * Returns the canonical name of the [ApiAnnotationsPlugin] class.
      */
     override val consumerId: String
         get() = ApiAnnotationsPlugin::class.java.canonicalName
@@ -62,9 +56,21 @@ internal abstract class Annotator<T>(
     }
 
     /**
+     * Tells if the given source file set is suitable for this renderer.
+     *
+     * The default implementation accepts all source file sets output directory
+     * of which ends with "java". Derived classes may override this method if they handle
+     * specific source code cases, e.g., generated by gRPC.
+     *
+     * @see <a href="https://github.com/SpineEventEngine/ProtoData/issues/150">ProtoData issue</a>
+     */
+    protected open fun suitableFor(sources: SourceFileSet): Boolean =
+        sources.outputRoot.endsWith("java")
+
+    /**
      * Obtains the annotation type by taking its name from [settings].
      */
-    protected fun annotationClass(apiOption: ApiOption): Class<Annotation> {
+    internal fun annotationClass(apiOption: ApiOption): Class<Annotation> {
         val annotationType = settings.annotationTypes
         val className = when (apiOption) {
             BETA -> annotationType.beta
@@ -78,31 +84,4 @@ internal abstract class Annotator<T>(
         val result = Class.forName(className) as Class<Annotation>
         return result
     }
-
-    /**
-     * Tells if the given source file set is suitable for this renderer.
-     *
-     * @see <a href="https://github.com/SpineEventEngine/ProtoData/issues/150">ProtoData issue</a>
-     */
-    protected open fun suitableFor(sources: SourceFileSet): Boolean =
-        sources.outputRoot.endsWith("java")
-
-    final override fun render(sources: SourceFileSet) {
-        if (suitableFor(sources)) {
-            this.sources = sources
-            doRender()
-        }
-    }
-
-    private fun doRender() {
-        val annotated: Set<T> = select(viewClass).all()
-        annotated.forEach {
-            annotate(it)
-        }
-    }
-
-    /**
-     * Annotates the code according to the given view state.
-     */
-    protected abstract fun annotate(view: T)
 }
